Added default values for keyPath parameters in the top level APIs.
By adding a default value for the keyPath parameter, it makes the public APIs more clear. It is confusing to add an empty string as the keyPath when you are trying to parse all the data. By adding a default value, the public APIs read much better and are more intuitive.